requirement.txt
protobuf==3.20.*
numpy
Pillow
mediapipe
opencv-python